refactor(structure): shell folders to features, components and pages
