+++ src/modules/Layout/Layout.component.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+const Layout = ({ children }) => (
+    <div>
+        {children}
+    </div>
+);
+
+/*
+    Layout propTypes
+*/
+Layout.propTypes = {
+    children: PropTypes.node,
+};
+Layout.defaultProps = {
+    children: null,
+};
+
+export default Layout;
