+++ src/pages/index.js
@@ -1,17 +1,16 @@
 import React from 'react';
-import logo from './logo.svg';
-import './App.css';
+
+import Layout from '../modules/Layout/Layout.component';
 
 const App = () => (
-    <div className="App">
+    <Layout>
         <header className="App-header">
-            <img src={logo} className="App-logo" alt="logo" />
             <h1 className="App-title">Welcome to React</h1>
         </header>
         <p className="App-intro">
             To get started, edit <code>src/App.js</code> and save to reload.
         </p>
-    </div>
+    </Layout>
 );
 
 export default App;
