+++ src/pages/index.test.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from './';
 
 it('renders without crashing', () => {
     const div = document.createElement('div');
